LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/position-try-tree-scoped-expected.txt


FAIL Document position-try-options matches @position-try in document scope assert_equals: expected 100 but got 999999
PASS Document position-try-options does not match @position-try in #outer_host scope
PASS Document position-try-options does not match @position-try in #inner_host scope
FAIL Outer position-try-options matches @position-try in document scope assert_equals: expected 100 but got 999999
FAIL Outer position-try-options matches @position-try in #outer_host scope assert_equals: expected 200 but got 999999
PASS Outer position-try-options does not match @position-try in #inner_host scope
FAIL Inner position-try-options matches @position-try in document scope assert_equals: expected 100 but got 999999
FAIL Inner position-try-options matches @position-try in #outer_host scope assert_equals: expected 200 but got 999999
FAIL Inner position-try-options matches @position-try in #inner_host scope assert_equals: expected 300 but got 999999
FAIL @position-try from same scope as :host rule assert_equals: expected 2 but got 999999
FAIL @position-try from same scope as ::slotted() rule assert_equals: expected 2 but got 999999
FAIL @position-try from same scope as ::part() rule assert_equals: expected 1 but got 999999
